Misc refactorings and nearly added melee attacks for the player.
The Animation recieved some major refactoring, things should make some more sense there now.

Changed the accessability levels of the variables in entity to be more restrictive.

An attack spritesheet for player is in, hooked up the animation in code, just need to finish the call to draw it.

Need to finish the logic so the melee attack only hits when mob is in range.